Mais implementações

+++ src/commonMain/kotlin/vitorscoelho/gyncanvas/core/event/CanvasEvent.kt
@@ -3,13 +3,13 @@ package vitorscoelho.gyncanvas.core.event
 interface CanvasEvent
 
 interface CanvasMouseEvent : CanvasEvent {
-    val x: Double
-    val y: Double
+    val x: Float
+    val y: Float
     val button: CanvasMouseButton
 }
 
 interface CanvasScrollEvent : CanvasEvent {
-    val x: Double
-    val y: Double
+    val x: Float
+    val y: Float
     val deltaY: Double
 }
